backend/requirements.txt
flask
flask-cors
openai
python-dotenv
jinja2
weasyprint
